functions/nas_bqload/requirements.txt
google-cloud-storage==1.13.2
google-cloud-bigquery==1.8.1
pandas==0.24.0
pandas-gbq==0.9.0
